# PERF: better perf on _ensure_data in core/algorithms, helping perf of unique, duplicated, factorize